requirements.txt
# This is compatible with Python 3.8.10
# Install with pip install -r requirements.txt
easydict==1.10
numpy==1.23.5
tqdm==4.64.1
torchvision==0.14.1
torch==1.13.1
opencv_python==4.6.0.66
Pillow==9.3.0
tensorboardX==2.5.1
pyserial